# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meridian Capital Portfolio Analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PortfolioSummary {
  /** Total AUM in USD billions */
  aum: number;
  aumChange1D: number;
  aumChangePct1D: number;
  /** NAV per unit */
  nav: number;
  navChange1D: number;
  navChangePct1D: number;
  /** Inception-to-date return % */
  totalReturn: number;
  ytdReturn: number;
  sharpeRatio: number;
  sortinoRatio: number;
  beta: number;
  /** Jensen's alpha annualised % */
  alpha: number;
  /** Maximum drawdown % */
  maxDrawdown: number;
  /** Annualised volatility % */
  volatility: number;
  cashWeight: number;
  investedWeight: number;
  lastUpdated: string;
  isStale: boolean;
}

export interface Holding {
  ticker: string;
  name: string;
  assetClass: string;
  sector: string;
  weight: number;
  marketValue: number;
  quantity: number;
  avgCost: number;
  currentPrice: number;
  dayReturn: number;
  dayReturnPct: number;
  totalReturn: number;
  totalReturnPct: number;
  beta: number;
  isStale: boolean;
}

export interface NavDataPoint {
  date: string;
  portfolioNav: number;
  benchmarkNav: number;
  /** Cumulative return % from period start */
  portfolioReturn: number;
  benchmarkReturn: number;
}

export interface NavSeriesResponse {
  period: string;
  benchmark: string;
  series: NavDataPoint[];
  lastUpdated: string;
  isStale: boolean;
}

export interface AllocationSlice {
  name: string;
  weight: number;
  benchmarkWeight: number;
  activeWeight: number;
  return1Y: number;
}

export interface RiskMetrics {
  sharpeRatio: number;
  sortinoRatio: number;
  calmarRatio: number;
  beta: number;
  alpha: number;
  rsquared: number;
  trackingError: number;
  informationRatio: number;
  volatility30D: number;
  volatility90D: number;
  volatilityAnnual: number;
  correlationToBenchmark: number;
  lastUpdated: string;
  isStale: boolean;
}

export interface VaRResult {
  /** VaR at 95% confidence in USD millions */
  var95: number;
  var99: number;
  /** CVaR/Expected Shortfall at 95% in USD millions */
  cvar95: number;
  cvar99: number;
  /** VaR as % of AUM */
  varPct95: number;
  varPct99: number;
  method: string;
  horizon: string;
  confidence: number;
  /** Array of simulated daily returns for histogram */
  returnDistribution: number[];
  lastUpdated: string;
  isStale: boolean;
}

export interface DrawdownDataPoint {
  date: string;
  /** Drawdown % (negative number, 0 at peaks) */
  drawdown: number;
  portfolioValue: number;
}

export interface DrawdownEvent {
  peakDate: string;
  troughDate: string;
  /** @nullable */
  recoveryDate: string | null;
  maxDrawdown: number;
  durationDays: number;
  /** @nullable */
  recoveryDays: number | null;
}

export interface DrawdownResponse {
  period: string;
  series: DrawdownDataPoint[];
  events: DrawdownEvent[];
  currentDrawdown: number;
  maxDrawdown: number;
  lastUpdated: string;
  isStale: boolean;
}

export interface CorrelationMatrixResponse {
  assets: string[];
  /** Row-major 2D correlation matrix (array of arrays) */
  matrix: number[][];
  lookbackDays: number;
  lastUpdated: string;
  isStale: boolean;
}

export interface BrinsonSector {
  sector: string;
  portfolioWeight: number;
  benchmarkWeight: number;
  portfolioReturn: number;
  benchmarkReturn: number;
  allocationEffect: number;
  selectionEffect: number;
  interactionEffect: number;
  totalActiveReturn: number;
}

export interface BrinsonResponse {
  period: string;
  sectors: BrinsonSector[];
  totalAllocationEffect: number;
  totalSelectionEffect: number;
  totalInteractionEffect: number;
  totalActiveReturn: number;
  lastUpdated: string;
  isStale: boolean;
}

export interface FactorExposure {
  /** Market | SMB | HML | MOM | QMJ | BAB */
  factor: string;
  /** Factor beta / loading */
  exposure: number;
  tstat: number;
  /** Contribution to portfolio return in % */
  contribution: number;
}

export interface StressScenario {
  name: string;
  /** historical | hypothetical */
  category: string;
  /** Estimated P&L in USD millions */
  estimatedLoss: number;
  estimatedLossPct: number;
  description: string;
}

export interface ActivityItem {
  id: string;
  /** TRADE | ALERT | THRESHOLD | REBALANCE | COMPLIANCE */
  type: string;
  message: string;
  timestamp: string;
  /** info | warning | critical */
  severity: string;
  /** @nullable */
  ticker?: string | null;
  /** @nullable */
  value?: number | null;
}

/**
 * Widget-specific settings (arbitrary JSON)
 */
export type WidgetConfigSettings = { [key: string]: unknown };

export interface WidgetConfig {
  id: string;
  /** PORTFOLIO_SUMMARY | NAV_CHART | VAR_GAUGE | DRAWDOWN | CORRELATION | ATTRIBUTION | HOLDINGS | ALLOCATION | RISK_METRICS | FACTORS | STRESS_TEST | ACTIVITY */
  type: string;
  x: number;
  y: number;
  w: number;
  h: number;
  /** Widget-specific settings (arbitrary JSON) */
  settings: WidgetConfigSettings;
}

export interface DashboardLayout {
  id: string;
  name: string;
  widgets: WidgetConfig[];
  /** Number of grid columns (12 | 16 | 24) */
  columns: number;
  createdAt: string;
  updatedAt: string;
}

export interface DashboardLayoutInput {
  name: string;
  widgets: WidgetConfig[];
  columns: number;
}

export interface ThemeConfig {
  id: string;
  name: string;
  /** bloomberg | refinitiv | midnight | arctic | meridian */
  preset: string;
}

export interface ThemeInput {
  preset: string;
}

export type GetHoldingsParams = {
assetClass?: string;
sector?: string;
};

export type GetNavSeriesParams = {
/**
 * 1M | 3M | 6M | 1Y | 3Y | 5Y | YTD
 */
period?: string;
/**
 * NIFTY50 | SP500 | MSCI_WORLD
 */
benchmark?: string;
};

export type GetValueAtRiskParams = {
/**
 * 0.90 | 0.95 | 0.99
 */
confidence?: string;
/**
 * 1D | 5D | 10D | 1M
 */
horizon?: string;
/**
 * historical | parametric
 */
method?: string;
};

export type GetDrawdownSeriesParams = {
period?: string;
};

export type GetCorrelationMatrixParams = {
/**
 * 30 | 60 | 90 | 252
 */
lookback?: number;
/**
 * Max number of assets to include
 */
limit?: number;
};

export type GetBrinsonAttributionParams = {
period?: string;
};

export type GetActivityFeedParams = {
limit?: number;
};