# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meridian Capital Portfolio Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portfolio/holdings:
    get:
      operationId: getHoldings
      tags: [portfolio]
      summary: Current holdings with weights, P&L, and sector tags
      parameters:
        - name: assetClass
          in: query
          schema:
            type: string
        - name: sector
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holding"

  # ── NAV Performance Chart (W03) ─────────────────────────────────────
...
  /risk/var:
    get:
      operationId: getValueAtRisk
      tags: [risk]
      summary: VaR and CVaR at multiple confidence levels and horizons
      parameters:
        - name: confidence
          in: query
          description: "0.90 | 0.95 | 0.99"
          schema:
            type: string
            default: "0.95"
        - name: horizon
          in: query
          description: "1D | 5D | 10D | 1M"
          schema:
            type: string
            default: "1D"
        - name: method
          in: query
          description: "historical | parametric"
          schema:
            type: string
            default: "historical"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaRResult"

  # ── Drawdown Analysis (W07) ─────────────────────────────────────────
  /risk/drawdown:
    get:
      operationId: getDrawdownSeries
      tags: [risk]
      summary: Drawdown time series with peak/trough annotations
      parameters:
        - name: period
          in: query
          schema:
            type: string
            default: "3Y"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawdownResponse"

  # ── Correlation Matrix (W08) ────────────────────────────────────────
  /risk/correlation:
    get:
      operationId: getCorrelationMatrix
      tags: [risk]
      summary: Pairwise correlation matrix for top holdings
      parameters:
        - name: lookback
          in: query
          description: "30 | 60 | 90 | 252"
          schema:
            type: integer
            default: 90
        - name: limit
          in: query
          description: Max number of assets to include
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrelationMatrixResponse"

  # ── Brinson Attribution (W09) ───────────────────────────────────────
...
  /portfolio/activity:
    get:
      operationId: getActivityFeed
      tags: [portfolio]
      summary: Recent trades, alerts, and threshold breaches
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ── Dashboard Layout Persistence ─────────────────────────────────────
...
components:
  schemas:
...
    Holding:
      type: object
      required:
        - ticker
        - name
        - assetClass
        - sector
        - weight
        - marketValue
        - quantity
        - avgCost
        - currentPrice
        - dayReturn
        - dayReturnPct
        - totalReturn
        - totalReturnPct
        - beta
        - isStale
      properties:
        ticker:
          type: string
        name:
          type: string
        assetClass:
          type: string
        sector:
          type: string
        weight:
          type: number
        marketValue:
          type: number
        quantity:
          type: number
        avgCost:
          type: number
        currentPrice:
          type: number
        dayReturn:
          type: number
        dayReturnPct:
          type: number
        totalReturn:
          type: number
        totalReturnPct:
          type: number
        beta:
          type: number
        isStale:
          type: boolean

    # ── NAV Performance ──────────────────────────────────────────────
...
    VaRResult:
      type: object
      required:
        - var95
        - var99
        - cvar95
        - cvar99
        - varPct95
        - varPct99
        - method
        - horizon
        - confidence
        - returnDistribution
        - lastUpdated
        - isStale
      properties:
        var95:
          type: number
          description: VaR at 95% confidence in USD millions
        var99:
          type: number
        cvar95:
          type: number
          description: CVaR/Expected Shortfall at 95% in USD millions
        cvar99:
          type: number
        varPct95:
          type: number
          description: VaR as % of AUM
        varPct99:
          type: number
        method:
          type: string
        horizon:
          type: string
        confidence:
          type: number
        returnDistribution:
          type: array
          items:
            type: number
          description: Array of simulated daily returns for histogram
        lastUpdated:
          type: string
          format: date-time
        isStale:
          type: boolean

    # ── Drawdown ──────────────────────────────────────────────────────
    DrawdownDataPoint:
      type: object
      required: [date, drawdown, portfolioValue]
      properties:
        date:
          type: string
          format: date
        drawdown:
          type: number
          description: Drawdown % (negative number, 0 at peaks)
        portfolioValue:
          type: number
    DrawdownEvent:
      type: object
      required: [peakDate, troughDate, recoveryDate, maxDrawdown, durationDays, recoveryDays]
      properties:
        peakDate:
          type: string
          format: date
        troughDate:
          type: string
          format: date
        recoveryDate:
          type: ["string", "null"]
          format: date
        maxDrawdown:
          type: number
        durationDays:
          type: integer
        recoveryDays:
          type: ["integer", "null"]
    DrawdownResponse:
      type: object
      required: [period, series, events, currentDrawdown, maxDrawdown, lastUpdated, isStale]
      properties:
        period:
          type: string
        series:
          type: array
          items:
            $ref: "#/components/schemas/DrawdownDataPoint"
        events:
          type: array
          items:
            $ref: "#/components/schemas/DrawdownEvent"
        currentDrawdown:
          type: number
        maxDrawdown:
          type: number
        lastUpdated:
          type: string
          format: date-time
        isStale:
          type: boolean

    # ── Correlation Matrix ────────────────────────────────────────────
    CorrelationMatrixResponse:
      type: object
      required: [assets, matrix, lookbackDays, lastUpdated, isStale]
      properties:
        assets:
          type: array
          items:
            type: string
        matrix:
          type: array
          description: Row-major 2D correlation matrix (array of arrays)
          items:
            type: array
            items:
              type: number
        lookbackDays:
          type: integer
        lastUpdated:
          type: string
          format: date-time
        isStale:
          type: boolean

    # ── Brinson Attribution ───────────────────────────────────────────
...
    ActivityItem:
      type: object
      required: [id, type, message, timestamp, severity]
      properties:
        id:
          type: string
        type:
          type: string
          description: "TRADE | ALERT | THRESHOLD | REBALANCE | COMPLIANCE"
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        severity:
          type: string
          description: "info | warning | critical"
        ticker:
          type: ["string", "null"]
        value:
          type: ["number", "null"]

    # ── Dashboard Layout ──────────────────────────────────────────────
...