# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meridian Capital Portfolio Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  /** TRADE | ALERT | THRESHOLD | REBALANCE | COMPLIANCE */
  type: string;
  message: string;
  timestamp: Date;
  /** info | warning | critical */
  severity: string;
  /** @nullable */
  ticker?: string | null;
  /** @nullable */
  value?: number | null;
}