# lib/api-zod/src/generated/types/holding.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meridian Capital Portfolio Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface Holding {
  ticker: string;
  name: string;
  assetClass: string;
  sector: string;
  weight: number;
  marketValue: number;
  quantity: number;
  avgCost: number;
  currentPrice: number;
  dayReturn: number;
  dayReturnPct: number;
  totalReturn: number;
  totalReturnPct: number;
  beta: number;
  isStale: boolean;
}