# lib/api-zod/src/generated/types/vaRResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meridian Capital Portfolio Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface VaRResult {
  /** VaR at 95% confidence in USD millions */
  var95: number;
  var99: number;
  /** CVaR/Expected Shortfall at 95% in USD millions */
  cvar95: number;
  cvar99: number;
  /** VaR as % of AUM */
  varPct95: number;
  varPct99: number;
  method: string;
  horizon: string;
  confidence: number;
  /** Array of simulated daily returns for histogram */
  returnDistribution: number[];
  lastUpdated: Date;
  isStale: boolean;
}